Move extract_enum_value into carina-core::utils to deduplicate
The inline `s.split('.').next_back().unwrap_or(s)` pattern in
carina-core/src/schema.rs and the standalone function in
carina-provider-aws/src/schemas/types.rs were identical. Move the
function into carina-core::utils alongside convert_enum_value, use it
in schema.rs enum validation, and update all carina-provider-aws call
sites to import from carina-core.

Closes #203
